change: Add google_cloud.query config field for rule-evaluator compatibility
See rationales in
https://github.com/GoogleCloudPlatform/prometheus-engine/pull/1711#discussion_r2254497026
